feat: models for club and club semester
