+++ packages/api/src/feature/club/service/club.public.service.ts
@@ -16,7 +16,7 @@ import { SemesterPublicService } from "@sparcs-clubs/api/feature/semester/public
 import UserPublicService from "@sparcs-clubs/api/feature/user/service/user.public.service";
 
 import { ClubDelegateDRepository } from "../delegate/club.club-delegate-d.repository";
-import { MClub } from "../model/club.model";
+import { MClubOld } from "../model/club-old.model";
 import ClubStudentTRepository from "../repository/club.club-student-t.repository";
 import ClubTRepository from "../repository/club.club-t.repository";
 import { DivisionPermanentClubDRepository } from "../repository/club.division-permanent-club-d.repository";
@@ -386,7 +386,7 @@ export default class ClubPublicService {
     clubId: number,
     semester: Pick<ISemester, "id"> | ISemester,
     date?: Date,
-  ): Promise<MClub> {
+  ): Promise<MClubOld> {
     const semesterParam =
       "endTerm" in semester
         ? semester
@@ -399,7 +399,7 @@ export default class ClubPublicService {
     clubId: number,
     semester: Pick<ISemester, "id"> | ISemester,
     date?: Date,
-  ): Promise<MClub | null> {
+  ): Promise<MClubOld | null> {
     const semesterParam =
       "endTerm" in semester
         ? semester
